update: extract file table row to component
